refactor: move embed edit container to image options
