+++ packages/blocks/src/page-block/default/components.ts
@@ -1,18 +1,7 @@
-import {
-  CaptionIcon,
-  CopyIcon,
-  DeleteIcon,
-  DownloadIcon,
-} from '@blocksuite/global/config';
 import { html } from 'lit';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import type { EditingState, IPoint } from '../../__internal__/index.js';
-import { tooltipStyle } from '../../components/tooltip/tooltip.js';
-import type { EmbedBlockModel } from '../../embed-block/embed-model.js';
-import { stopPropagation } from '../edgeless/utils.js';
-import type { DefaultSelectionSlots } from './default-page-block.js';
-import { copyImage, downloadImage, focusCaption } from './utils.js';
+import type { IPoint } from '../../__internal__/index.js';
 
 export function DraggingArea(rect: DOMRect | null) {
   if (rect === null) return null;
@@ -42,11 +31,14 @@ export function EmbedSelectedRectsContainer(
 ) {
   return html`
     <style>
-      .affine-page-selected-embed-rects-container > div {
+      .affine-page-selected-embed-rects-container > .resizes {
         position: absolute;
         display: block;
         border: 2px solid var(--affine-primary-color);
         user-select: none;
+      }
+      .affine-page-selected-embed-rects-container .resize {
+        pointer-events: auto;
       }
     </style>
     <div class="affine-page-selected-embed-rects-container resizable">
@@ -56,6 +48,7 @@ export function EmbedSelectedRectsContainer(
           top: rect.top + viewportOffset.y + 'px',
           width: rect.width + 'px',
           height: rect.height + 'px',
+          pointerEvents: 'none',
         };
         return html`
           <div class="resizes" style=${styleMap(style)}>
@@ -66,88 +59,6 @@ export function EmbedSelectedRectsContainer(
           </div>
         `;
       })}
-    </div>
-  `;
-}
-
-export function EmbedEditingContainer(
-  embedEditingState: EditingState | null,
-  slots: DefaultSelectionSlots,
-  viewportOffset: IPoint
-) {
-  if (!embedEditingState) return null;
-
-  const {
-    rect: { x, y },
-    model,
-  } = embedEditingState;
-  const style = {
-    left: x + viewportOffset.x + 'px',
-    top: y + viewportOffset.y + 'px',
-  };
-
-  return html`
-    <style>
-      .affine-embed-editing-state-container > div {
-        position: absolute;
-        display: block;
-        z-index: 1;
-      }
-      ${tooltipStyle}
-    </style>
-
-    <div
-      class="affine-embed-editing-state-container"
-      @pointerdown=${stopPropagation}
-    >
-      <div style=${styleMap(style)} class="embed-editing-state">
-        <format-bar-button
-          class="has-tool-tip"
-          width="100%"
-          @click=${() => {
-            focusCaption(model);
-            slots.embedRectsUpdated.emit([]);
-          }}
-        >
-          ${CaptionIcon}
-          <tool-tip inert tip-position="right" role="tooltip">Caption</tool-tip>
-        </format-bar-button>
-        <format-bar-button
-          class="has-tool-tip"
-          width="100%"
-          @click=${() => {
-            downloadImage(model);
-          }}
-        >
-          ${DownloadIcon}
-          <tool-tip inert tip-position="right" role="tooltip"
-            >Download
-          </tool-tip>
-        </format-bar-button>
-        <format-bar-button
-          class="has-tool-tip"
-          width="100%"
-          @click=${() => {
-            copyImage(model as EmbedBlockModel);
-          }}
-        >
-          ${CopyIcon}
-          <tool-tip inert tip-position="right" role="tooltip"
-            >Copy to clipboard
-          </tool-tip>
-        </format-bar-button>
-        <format-bar-button
-          class="has-tool-tip"
-          width="100%"
-          @click="${() => {
-            model.page.deleteBlock(model);
-            slots.embedRectsUpdated.emit([]);
-          }}"
-        >
-          ${DeleteIcon}
-          <tool-tip inert tip-position="right" role="tooltip">Delete</tool-tip>
-        </format-bar-button>
-      </div>
     </div>
   `;
 }
